Moved model to the right folder
