+++ backend/src/main/java/nl/tudelft/thefirstorder/spring/models/players/Player.java
@@ -1,4 +1,4 @@
-package nl.tudelft.thefirstorder.model.players;
+package nl.tudelft.thefirstorder.spring.models.players;
 
 /**
  * The Player class represents a player and his/her location on the map.
